Edit building scenes doc
review and merge

Question:
At one point we say "For simple models that have no animations, OBJ is still a common and reliable choice." but then we say that the Obj interface is going to be removed in the future. Should we remove that first mention? Or are we saying that what will be deprecated is just the "obj-model` interface, but that there will still be other ways to use OBJ models?

I also feel like it's worth adding something about creating custom interfaces for creating entities. I want to add that in the future.