Update 2021-05-31-wearables-editor-user-guide.md
Just changed this to clarify that there are in fact royalites now:
### Secondary sales

Items can be sold in secondary sales at any time, and for any price, in the Decentraland Marketplace only after:

- They have been **minted**, or
- they have been **purchased in a primary sale**

In other words, anybody who owns an NFT for a wearable can sell it in the Decentraland Marketplace. The original owner of the NFT will recieve 2.5% as royalties for each sale in the secondary market.

This is what it sayed before:
In other words, anybody who owns an NFT for a wearable can sell it in the Decentraland Marketplace. There are no royalties for wearables sold in secondary sales in Decentraland.